fix(entity): 6 类实体元数据 wire 字段类型对齐 vanilla 1.21.11 修真客户端类型校验崩
彻查并深修除 Wolf(已提交 1f9be27ca)外 6 类实体的 SynchedEntityData wire 字段类型/缺失错位
致真 Java 客户端 set_entity_data 严格类型校验崩溃问题:

- Bee: ANGER_TIME i32→i64(Long, serializerId 2) 对齐 vanilla Bee.DATA_ANGER_END_TIME
- Enderman: DATA_CARRIED_BLOCK_STATE i32→OptionalBlockStateValue(Optional<BlockState>,
  serializerId 15) 对齐 vanilla EnderMan.DATA_CARRY_STATE
- ZombieVillager: 3 个独立 i32(VILLAGER_TYPE/PROFESSION/LEVEL) 收敛为单一复合
  VillagerDataValue(VILLAGER_DATA, serializerId 18),wire id 19..22→19..20 对齐
  vanilla DATA_VILLAGER_DATA
- Sniffer: DATA_STATE i8(Byte)→SnifferStateValue(新增, serializerId 31 SNIFFER_STATE,
  VarInt(State.id)) 对齐 vanilla Sniffer.DATA_STATE;扩展 DataValue variant index 22 +
  EntityMetadataSerializer 枚举/编解码
- Cat: 补 DATA_VARIANT_ID(HolderVariantValue→serializerId 21 CAT_VARIANT, wire 19) 与
  DATA_COLLAR_COLOR(i32, wire 22) 两个缺失字段,修复旧实现仅 lying/relax 两字段致
  wire 19 错发 Boolean 撞客户端期望 CatVariant 崩;CatType 枚举 0-10 与 vanilla
  CAT_VARIANT 注册表 id 一一对应
- Minecart: 删 rolling_amp/rolling_dir/show_block 三个 ghost 字段(vanilla 无),
  display_tile i32→OptionalBlockStateValue 修 wire 11 类型崩,补 HURT/HURTDIR;
  5 字段对齐 vanilla wire 8-12;摇晃动画改本地成员驱动 + TODO 待接 EntityEvent
  (渲染回退,客户端暂看不到矿车受损摇晃)

配套: ClientEntity Cat 分支补 variant/collar TODO 桩;isAngry 标注预存在占位 bug;
golden test 加 Cat/Minecart 类型断言 + 更新 ZombieVillager id 范围;README 文档同步。

+++ src/common/entity/entities/passive/special/BeeEntity.cpp
@@ -64,7 +64,7 @@ namespace mc {
 // ============================================================================
 
 entity::DataParameter<i8> BeeEntity::DATA_FLAGS_PARAM = entity::EntityDataManager::createKey<i8>();
-entity::DataParameter<i32> BeeEntity::ANGER_TIME_PARAM = entity::EntityDataManager::createKey<i32>();
+entity::DataParameter<i64> BeeEntity::ANGER_TIME_PARAM = entity::EntityDataManager::createKey<i64>();
 
 // ============================================================================
 // 继承链标识（复刻 vanilla ClassTreeIdRegistry，parent = AnimalEntity::classInfo()）
@@ -128,7 +128,9 @@ void BeeEntity::registerData()
     entity::EntityDataManager::ClassRegisterGuard guard(m_dataManager, classInfo());
 
     m_dataManager.registerParam(DATA_FLAGS_PARAM, static_cast<i8>(0));
-    m_dataManager.registerParam(ANGER_TIME_PARAM, static_cast<i32>(0));
+    // ANGER_TIME 以 i64(Long) 注册对齐 vanilla Bee.DATA_ANGER_END_TIME(serializerId=2)。
+    // 业务层以 i32 语义运转,默认 0=非愤怒。旧 i32 致真客户端 field 类型校验崩。
+    m_dataManager.registerParam(ANGER_TIME_PARAM, static_cast<i64>(0));
 }
 
 // ============================================================================
@@ -186,13 +188,15 @@ void BeeEntity::setHasStung(bool stung)
 
 i32 BeeEntity::getAngerTime() const
 {
-    return m_dataManager.get(ANGER_TIME_PARAM);
+    // wire 以 i64(Long) 存储(对齐 vanilla),业务层返回 i32
+    return static_cast<i32>(m_dataManager.get(ANGER_TIME_PARAM));
 }
 
 void BeeEntity::setAngerTime(i32 time)
 {
     m_angerTime = time;
-    m_dataManager.set(ANGER_TIME_PARAM, time);
+    // 写入 i64(Long) 对齐 vanilla wire 类型(serializerId=2 VAR_LONG)
+    m_dataManager.set(ANGER_TIME_PARAM, static_cast<i64>(time));
 }
 
 void BeeEntity::setAngry(bool angry)
